tests/test_user_customization.py
```python
import unittest
from src.user_customization import customizeResponse

class TestUserCustomization(unittest.TestCase):

    def setUp(self):
        self.user_preferences = {
            "tone": "formal",
            "style": "verbose"
        }

    def test_customizeResponse(self):
        response = "Hello, how can I assist you today?"
        customized_response = customizeResponse(response, self.user_preferences)
        self.assertIsInstance(customized_response, str)

    def test_customizeResponse_tone(self):
        response = "Hello, how can I assist you today?"
        self.user_preferences["tone"] = "casual"
        customized_response = customizeResponse(response, self.user_preferences)
        self.assertIn("hey", customized_response.lower())

    def test_customizeResponse_style(self):
        response = "Hello, how can I assist you today?"
        self.user_preferences["style"] = "concise"
        customized_response = customizeResponse(response, self.user_preferences)
        self.assertTrue(len(customized_response.split()) <= len(response.split()))

if __name__ == '__main__':
    unittest.main()
```